+++ CHANGELOG.md
@@ -7,6 +7,9 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 ## [Unreleased]
 
+### Fixed
+- IntelliJ plugin crashing on run because both include and exclude were being set for env vars.
+
 ## 3.1.0
 
 ### Added